start/stop
added a start/stop button. I fixed an unknown bug in the process? I
think that each time a new image was loaded the interval was set so we
ended up with many intervals running.

Next up: space bar to also toggle processing, try using a interval for
drawing and an interval for processing, button to show original image,
buttons to increase and decrease processing speed, sound to go along
with it