Spacebar callback to toggleProcessing
spacebar calls back to toggleProcessing when it is not being used in an
edit text fashion. In this case it's default behavior is also overriden.

Next up: try using a interval for
drawing and an interval for processing, button to show original image,
buttons to increase and decrease processing speed, sound to go along
with it